split-style-and-metadata moved to helpers.js
